requirements.txt
discord.py
aiohttp
beautifulsoup4
rapidfuzz
playwright